Removed unused emails.ts and improve names of routes files
