+++ src/routes/user.ts
@@ -18,6 +18,7 @@ import {
 } from "../queries/query";
 import { User, PasswordResetToken } from "../queries/database";
 import { hashPassword, generateRandomToken } from "../queries/helper";
+import { LoggedInRequest } from "./request_types";
 
 const RESET_PASSWORD_EXPIRY_HOURS = 24;
 
@@ -119,19 +120,11 @@ async function loginUser(
   return h.response({ message: errorMessage }).code(401);
 }
 
-type UserDetailsRequest = Request & {
-  auth: {
-    credentials: {
-      user_id: number;
-    };
-  };
-};
-
 /**
  * Return the details of authenticated user
  */
 async function getAuthUserDetails(
-  request: UserDetailsRequest,
+  request: LoggedInRequest,
   h: ResponseToolkit,
   d: any
 ): Promise<ResponseObject> {
@@ -172,7 +165,7 @@ async function getAuthUserDetails(
  * Update the email of autheticated user
  */
 async function changeEmail(
-  request: Request,
+  request: LoggedInRequest,
   h: ResponseToolkit,
   d: any
 ): Promise<ResponseObject> {
@@ -201,7 +194,7 @@ async function changeEmail(
  * Change the user detail of the authenticated user
  */
 async function changeUserDetail(
-  request: Request,
+  request: LoggedInRequest,
   h: ResponseToolkit,
   d: any
 ): Promise<ResponseObject> {
@@ -237,7 +230,7 @@ async function changeUserDetail(
   return h.response().code(200);
 }
 
-type ChangePasswordRequest = Request & {
+type ChangePasswordRequest = LoggedInRequest & {
   payload: {
     password: string;
   };
@@ -327,17 +320,12 @@ async function resetPassword(
   return h.response().code(200);
 }
 
-type UserFeedbackRequest = Request & {
+type UserFeedbackRequest = LoggedInRequest & {
   payload: {
     question1: string;
     question2: string;
     question3: string;
     question4: string;
-  };
-  auth: {
-    credentials: {
-      user_id: number;
-    };
   };
 };
 
@@ -414,7 +402,7 @@ async function updateAskForFeedback(
   return h.response().code(200);
 }
 
-/** 
+/**
  * Check the user feedback flag
  * This is used to determine if the user should be asked for feedback
  */
@@ -438,7 +426,7 @@ async function getUserAskForFeedback(
   return h.response({ askForFeedback }).code(200);
 }
 
-export const databaseRoutes: ServerRoute[] = [
+export const userRoutes: ServerRoute[] = [
   /** Public APIs */
   // Register a new account
   {
